requirements.txt
PyGithub>=2.1.1
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
openai>=1.30.0
python-dotenv>=1.0.0
python-dotenv>=1.0
pytest>=8.0

